Make an endpoints proxy struct
