+++ demos/nrf52840/nrf52840-null/src/main.rs
@@ -105,13 +105,13 @@ async fn button_worker(mut btn: Input<'static>, name: &'static str) {
             continue;
         }
         STACK
-            .req_resp::<LedEndpoint>(Address::unknown(), &true, Some(name))
+            .endpoints().request::<LedEndpoint>(Address::unknown(), &true, Some(name))
             .await
             .unwrap();
         let _ = STACK.broadcast_topic::<ButtonPressedTopic>(&1, None);
         btn.wait_for_high().await;
         STACK
-            .req_resp::<LedEndpoint>(Address::unknown(), &false, Some(name))
+            .endpoints().request::<LedEndpoint>(Address::unknown(), &false, Some(name))
             .await
             .unwrap();
     }
