+++ demos/rp2040/rp2040-serial-pair/src/bin/target.rs
@@ -74,11 +74,11 @@ async fn main(spawner: Spawner) {
     loop {
         ticker.next().await;
         let _ = STACK
-            .req_resp::<LedEndpoint>(Address::unknown(), &true, Some("led"))
+            .endpoints().request::<LedEndpoint>(Address::unknown(), &true, Some("led"))
             .await;
         ticker.next().await;
         let _ = STACK
-            .req_resp::<LedEndpoint>(Address::unknown(), &false, Some("led"))
+            .endpoints().request::<LedEndpoint>(Address::unknown(), &false, Some("led"))
             .await;
     }
 }
@@ -90,7 +90,7 @@ async fn pinger() {
     loop {
         ticker.next().await;
         let res = STACK
-            .req_resp::<ErgotPingEndpoint>(
+            .endpoints().request::<ErgotPingEndpoint>(
                 Address {
                     network_id: 1,
                     node_id: 1,
